pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Родион Гнутов",email = "rodion.gnutov.2000@mail.ru"}
]

requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "httpie (>=3.2.4,<4.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[tool.poetry]
packages = [{include = "app", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
